Rename reciever -> receiver
